requirements.txt
aiogram==3.0.0b7
SQLAlchemy
aiosqlite==0.17.0
python-dotenv==1.0.0
uvicorn==0.18.3